Reorganize LSP and VS Code extension code structure
Split monolithic lsp/server.go (1316 lines, 47 functions) into focused
files: dispatch.go (parse routing + error conversion), handler.go (thin
LSP stubs), gsl.go (GSL-specific features), gql.go (GQL-specific
features), helpers.go (shared utilities). server.go now contains only
the Server struct, lifecycle, and document management (~200 lines).

Move VS Code extension from gitignored vscode-gsl/ to editors/vscode/
following the convention used by rust-analyzer and other multi-editor
projects. Fix findServerBinary() path for the new location (2 levels
up instead of 1). Update .gitignore accordingly.
