+++ editors/vscode/.vscodeignore
@@ -1,0 +1,4 @@
+node_modules/
+.git/
+.gitignore
+*.vsix
