+++ lsp/server.go
@@ -2,12 +2,6 @@ package lsp
 
 import (
 	"context"
-	"fmt"
-	"log"
-	"regexp"
-	"sort"
-	"strconv"
-	"strings"
 	"sync"
 
 	"github.com/dnnrly/gsl-lang"
@@ -148,333 +142,6 @@ func (s *Server) DidClose(ctx context.Context, params *protocol.DidCloseTextDocu
 	return nil
 }
 
-func isGQLFile(docURI uri.URI) bool {
-	return strings.HasSuffix(strings.ToLower(string(docURI)), ".gql")
-}
-
-func (s *Server) parseAndDiagnose(ctx context.Context, doc *document) {
-	if isGQLFile(doc.docURI) {
-		log.Printf("gsl-lsp: parsing GQL file %s", doc.docURI)
-		s.parseGQLAndDiagnose(ctx, doc)
-		return
-	}
-	log.Printf("gsl-lsp: parsing GSL file %s", doc.docURI)
-
-	graph, pErr := gsl.Parse(strings.NewReader(doc.content))
-	doc.graph = graph
-
-	var diags []protocol.Diagnostic
-
-	if pErr != nil && pErr.HasError() {
-		d := parseErrorToDiagnostic(pErr.Err.Error(), doc.content)
-		if d != nil {
-			diags = append(diags, *d)
-		} else {
-			diags = append(diags, protocol.Diagnostic{
-				Range: protocol.Range{
-					Start: protocol.Position{Line: 0, Character: 0},
-					End:   protocol.Position{Line: 0, Character: 0},
-				},
-				Severity: protocol.DiagnosticSeverityError,
-				Source:   protocol.NewOptional("gsl"),
-				Message:  protocol.String(pErr.Err.Error()),
-			})
-		}
-	}
-
-	if pErr != nil && pErr.HasWarnings() {
-		for _, w := range pErr.Warnings {
-			d := parseErrorToDiagnostic(w.Error(), doc.content)
-			if d != nil {
-				d.Severity = protocol.DiagnosticSeverityWarning
-				diags = append(diags, *d)
-			} else {
-				diags = append(diags, protocol.Diagnostic{
-					Range:    protocol.Range{Start: protocol.Position{Line: 0, Character: 0}, End: protocol.Position{Line: 0, Character: 0}},
-					Severity: protocol.DiagnosticSeverityWarning,
-					Source:   protocol.NewOptional("gsl"),
-					Message:  protocol.String(w.Error()),
-				})
-			}
-		}
-	}
-
-	doc.pErrs = diags
-
-	_ = s.client.PublishDiagnostics(ctx, &protocol.PublishDiagnosticsParams{
-		URI:         doc.docURI,
-		Version:     protocol.NewOptional(doc.version),
-		Diagnostics: diags,
-	})
-}
-
-func (s *Server) parseGQLAndDiagnose(ctx context.Context, doc *document) {
-	parser := query.NewQueryParser(doc.content)
-	q, err := parser.Parse()
-	doc.gqlQuery = q
-	doc.gqlErr = err
-
-	var diags []protocol.Diagnostic
-	if err != nil {
-		d := gqlParseErrorToDiagnostic(err, doc.content)
-		if d != nil {
-			diags = append(diags, *d)
-		} else {
-			diags = append(diags, protocol.Diagnostic{
-				Range: protocol.Range{
-					Start: protocol.Position{Line: 0, Character: 0},
-					End:   protocol.Position{Line: 0, Character: 0},
-				},
-				Severity: protocol.DiagnosticSeverityError,
-				Source:   protocol.NewOptional("gql"),
-				Message:  protocol.String(err.Error()),
-			})
-		}
-	}
-	doc.pErrs = diags
-
-	_ = s.client.PublishDiagnostics(ctx, &protocol.PublishDiagnosticsParams{
-		URI:         doc.docURI,
-		Version:     protocol.NewOptional(doc.version),
-		Diagnostics: diags,
-	})
-}
-
-var errPosRe = regexp.MustCompile(`at (\d+):(\d+)$`)
-var gqlErrPosRe = regexp.MustCompile(`(\d+):(\d+):`)
-
-func parseErrorToDiagnostic(msg, content string) *protocol.Diagnostic {
-	m := errPosRe.FindStringSubmatch(msg)
-	if m == nil {
-		return nil
-	}
-
-	line, _ := strconv.Atoi(m[1])
-	col, _ := strconv.Atoi(m[2])
-
-	start := protocol.Position{
-		Line:      uint32(max(0, line-1)),
-		Character: uint32(max(0, col-1)),
-	}
-
-	lines := strings.Split(content, "\n")
-	end := start
-	if int(start.Line) < len(lines) {
-		lineStr := lines[start.Line]
-		end.Character = uint32(len(lineStr))
-	}
-
-	cleanMsg := errPosRe.ReplaceAllString(msg, "")
-	cleanMsg = strings.TrimSpace(cleanMsg)
-
-	return &protocol.Diagnostic{
-		Range:    protocol.Range{Start: start, End: end},
-		Severity: protocol.DiagnosticSeverityError,
-		Source:   protocol.NewOptional("gsl"),
-		Message:  protocol.String(cleanMsg),
-	}
-}
-
-func gqlParseErrorToDiagnostic(err error, content string) *protocol.Diagnostic {
-	msg := err.Error()
-	m := gqlErrPosRe.FindStringSubmatch(msg)
-	if m != nil {
-		line, _ := strconv.Atoi(m[1])
-		col, _ := strconv.Atoi(m[2])
-		start := protocol.Position{
-			Line:      uint32(max(0, line-1)),
-			Character: uint32(max(0, col-1)),
-		}
-		lines := strings.Split(content, "\n")
-		end := start
-		if int(start.Line) < len(lines) {
-			end.Character = uint32(len(lines[start.Line]))
-		}
-
-		cleanMsg := gqlErrPosRe.ReplaceAllString(msg, "")
-		cleanMsg = strings.TrimSpace(cleanMsg)
-
-		return &protocol.Diagnostic{
-			Range:    protocol.Range{Start: start, End: end},
-			Severity: protocol.DiagnosticSeverityError,
-			Source:   protocol.NewOptional("gql"),
-			Message:  protocol.String(cleanMsg),
-		}
-	}
-
-	return nil
-}
-
-func (s *Server) Completion(ctx context.Context, params *protocol.CompletionParams) (protocol.CompletionResult, error) {
-	s.mu.Lock()
-	doc, ok := s.documents[params.TextDocument.URI]
-	s.mu.Unlock()
-	if !ok {
-		return protocol.CompletionItemSlice{}, nil
-	}
-
-	if isGQLFile(doc.docURI) {
-		items := completeGQL(doc.content, int(params.Position.Line), int(params.Position.Character))
-		return protocol.CompletionItemSlice(items), nil
-	}
-
-	pos := params.Position
-	items := s.complete(doc, int(pos.Line), int(pos.Character))
-	return protocol.CompletionItemSlice(items), nil
-}
-
-func (s *Server) Hover(ctx context.Context, params *protocol.HoverParams) (*protocol.Hover, error) {
-	s.mu.Lock()
-	doc, ok := s.documents[params.TextDocument.URI]
-	s.mu.Unlock()
-	if !ok {
-		return nil, nil
-	}
-
-	if isGQLFile(doc.docURI) {
-		return hoverGQL(doc, int(params.Position.Line), int(params.Position.Character)), nil
-	}
-
-	if doc.graph == nil {
-		return nil, nil
-	}
-
-	pos := params.Position
-	word := wordAt(doc.content, int(pos.Line), int(pos.Character))
-	if word == nil {
-		return nil, nil
-	}
-
-	info := s.hoverFor(doc.graph, word.word)
-	if info == "" {
-		return nil, nil
-	}
-
-	rnj := protocol.Range{
-		Start: protocol.Position{Line: uint32(word.line), Character: uint32(word.colS)},
-		End:   protocol.Position{Line: uint32(word.line), Character: uint32(word.colE)},
-	}
-
-	return &protocol.Hover{
-		Contents: &protocol.MarkupContent{
-			Kind:  protocol.MarkupKindMarkdown,
-			Value: info,
-		},
-		Range: &rnj,
-	}, nil
-}
-
-func (s *Server) Definition(ctx context.Context, params *protocol.DefinitionParams) (protocol.DefinitionResult, error) {
-	s.mu.Lock()
-	doc, ok := s.documents[params.TextDocument.URI]
-	s.mu.Unlock()
-	if !ok || isGQLFile(doc.docURI) {
-		return nil, nil
-	}
-
-	pos := params.Position
-	word := wordAt(doc.content, int(pos.Line), int(pos.Character))
-	if word == nil {
-		return nil, nil
-	}
-
-	tokens := tokenize(doc.content)
-	loc := findDeclPosition(tokens, word.word, doc.docURI)
-	if loc == nil {
-		return nil, nil
-	}
-	return loc, nil
-}
-
-func (s *Server) References(ctx context.Context, params *protocol.ReferenceParams) ([]protocol.Location, error) {
-	s.mu.Lock()
-	doc, ok := s.documents[params.TextDocument.URI]
-	s.mu.Unlock()
-	if !ok || isGQLFile(doc.docURI) {
-		return nil, nil
-	}
-
-	pos := params.Position
-	word := wordAt(doc.content, int(pos.Line), int(pos.Character))
-	if word == nil {
-		return nil, nil
-	}
-
-	tokens := tokenize(doc.content)
-	return findReferences(tokens, word.word, doc.docURI), nil
-}
-
-func (s *Server) DocumentSymbol(ctx context.Context, params *protocol.DocumentSymbolParams) (protocol.DocumentSymbolResult, error) {
-	s.mu.Lock()
-	doc, ok := s.documents[params.TextDocument.URI]
-	s.mu.Unlock()
-	if !ok {
-		return protocol.DocumentSymbolSlice{}, nil
-	}
-
-	if isGQLFile(doc.docURI) {
-		symbols := gqlDocumentSymbols(doc.content)
-		return protocol.DocumentSymbolSlice(symbols), nil
-	}
-
-	tokens := tokenize(doc.content)
-	symbols := buildSymbols(tokens)
-	return protocol.DocumentSymbolSlice(symbols), nil
-}
-
-func (s *Server) Formatting(ctx context.Context, params *protocol.DocumentFormattingParams) ([]protocol.TextEdit, error) {
-	s.mu.Lock()
-	doc, ok := s.documents[params.TextDocument.URI]
-	s.mu.Unlock()
-	if !ok || doc.graph == nil || isGQLFile(doc.docURI) {
-		return nil, nil
-	}
-
-	formatted := gsl.Serialize(doc.graph)
-	lines := strings.Split(doc.content, "\n")
-	lastLine := uint32(max(0, len(lines)-1))
-	lastChar := uint32(0)
-	if len(lines) > 0 {
-		lastChar = uint32(len(lines[len(lines)-1]))
-	}
-
-	return []protocol.TextEdit{
-		{
-			Range: protocol.Range{
-				Start: protocol.Position{Line: 0, Character: 0},
-				End:   protocol.Position{Line: lastLine, Character: lastChar},
-			},
-			NewText: formatted,
-		},
-	}, nil
-}
-
-func (s *Server) SemanticTokensFull(ctx context.Context, params *protocol.SemanticTokensParams) (*protocol.SemanticTokens, error) {
-	s.mu.Lock()
-	doc, ok := s.documents[params.TextDocument.URI]
-	s.mu.Unlock()
-	if !ok {
-		return &protocol.SemanticTokens{Data: []uint32{}}, nil
-	}
-
-	if isGQLFile(doc.docURI) {
-		data := gqlSemanticTokens(doc.content)
-		return &protocol.SemanticTokens{Data: data}, nil
-	}
-
-	tokens := tokenize(doc.content)
-	data := encodeSemanticTokens(tokens)
-	return &protocol.SemanticTokens{Data: data}, nil
-}
-
-func (s *Server) DidChangeConfiguration(ctx context.Context, params *protocol.DidChangeConfigurationParams) error {
-	return nil
-}
-
-func (s *Server) DidChangeWatchedFiles(ctx context.Context, params *protocol.DidChangeWatchedFilesParams) error {
-	return nil
-}
-
 func (s *Server) DidSave(ctx context.Context, params *protocol.DidSaveTextDocumentParams) error {
 	s.mu.Lock()
 	defer s.mu.Unlock()
@@ -486,831 +153,10 @@ func (s *Server) DidSave(ctx context.Context, params *protocol.DidSaveTextDocume
 	return nil
 }
 
-var gqlKeywords = []struct {
-	label  string
-	detail string
-	kind   protocol.CompletionItemKind
-}{
-	{"subgraph", "Filter nodes/edges by predicate", protocol.CompletionItemKindKeyword},
-	{"from", "Select input graph", protocol.CompletionItemKindKeyword},
-	{"make", "Set attributes matching predicate", protocol.CompletionItemKindKeyword},
-	{"remove", "Remove edges, attributes, or orphans", protocol.CompletionItemKindKeyword},
-	{"collapse", "Merge nodes matching predicate", protocol.CompletionItemKindKeyword},
-	{"into", "Target node ID for collapse", protocol.CompletionItemKindKeyword},
-	{"where", "Predicate clause", protocol.CompletionItemKindKeyword},
-	{"traverse", "Graph traversal", protocol.CompletionItemKindKeyword},
-	{"scope", "Sugar for traverse down all", protocol.CompletionItemKindKeyword},
-	{"node", "Element type for predicate", protocol.CompletionItemKindKeyword},
-	{"edge", "Element type for predicate", protocol.CompletionItemKindKeyword},
-	{"exists", "Attribute existence check", protocol.CompletionItemKindKeyword},
-	{"not", "Negation", protocol.CompletionItemKindKeyword},
-	{"AND", "Logical AND", protocol.CompletionItemKindKeyword},
-	{"in", "Set membership / traversal direction", protocol.CompletionItemKindKeyword},
-	{"out", "Traversal direction", protocol.CompletionItemKindKeyword},
-	{"both", "Traversal direction", protocol.CompletionItemKindKeyword},
-	{"up", "Traversal direction (dependency)", protocol.CompletionItemKindKeyword},
-	{"down", "Traversal direction (dependency)", protocol.CompletionItemKindKeyword},
-	{"all", "Unlimited traversal depth", protocol.CompletionItemKindKeyword},
-	{"orphans", "Remove isolated nodes", protocol.CompletionItemKindKeyword},
-	{"parent", "Edge parent predicate", protocol.CompletionItemKindKeyword},
-	{"depth", "Edge dependency depth", protocol.CompletionItemKindKeyword},
-	{"depends", "Edge dependency predicate", protocol.CompletionItemKindKeyword},
-	{"on", "Part of 'depends on'", protocol.CompletionItemKindKeyword},
-	{"true", "Boolean true", protocol.CompletionItemKindKeyword},
-	{"false", "Boolean false", protocol.CompletionItemKindKeyword},
-	{"as", "Named graph binding", protocol.CompletionItemKindKeyword},
-}
-
-func completeGQL(content string, line, character int) []protocol.CompletionItem {
-	lines := strings.Split(content, "\n")
-	if line >= len(lines) {
-		return keywordCompletionsGQL("")
-	}
-	lineStr := lines[line]
-	prefix := lineStr[:min(character, len(lineStr))]
-	trimmed := strings.TrimSpace(prefix)
-
-	lastWord := trimmed
-	if idx := strings.LastIndexAny(trimmed, " \t|("); idx >= 0 {
-		lastWord = trimmed[idx+1:]
-	}
-
-	return keywordCompletionsGQL(lastWord)
-}
-
-func keywordCompletionsGQL(prefix string) []protocol.CompletionItem {
-	var items []protocol.CompletionItem
-	for _, kw := range gqlKeywords {
-		if prefix == "" || strings.HasPrefix(kw.label, prefix) {
-			items = append(items, protocol.CompletionItem{
-				Label:      kw.label,
-				Detail:     protocol.NewOptional(kw.detail),
-				Kind:       kw.kind,
-				InsertText: protocol.NewOptional(kw.label),
-			})
-		}
-	}
-	return items
-}
-
-func hoverGQL(doc *document, line, character int) *protocol.Hover {
-	word := wordAt(doc.content, line, character)
-	if word == nil {
-		return nil
-	}
-
-	for _, kw := range gqlKeywords {
-		if kw.label == word.word {
-			rnj := protocol.Range{
-				Start: protocol.Position{Line: uint32(word.line), Character: uint32(word.colS)},
-				End:   protocol.Position{Line: uint32(word.line), Character: uint32(word.colE)},
-			}
-			return &protocol.Hover{
-				Contents: &protocol.MarkupContent{
-					Kind:  protocol.MarkupKindMarkdown,
-					Value: fmt.Sprintf("**`%s`** — %s", kw.label, kw.detail),
-				},
-				Range: &rnj,
-			}
-		}
-	}
+func (s *Server) DidChangeConfiguration(ctx context.Context, params *protocol.DidChangeConfigurationParams) error {
 	return nil
 }
 
-func gqlDocumentSymbols(content string) []protocol.DocumentSymbol {
-	lines := strings.Split(content, "\n")
-	var symbols []protocol.DocumentSymbol
-
-	for i, line := range lines {
-		trimmed := strings.TrimSpace(line)
-		if trimmed == "" || strings.HasPrefix(trimmed, "#") {
-			continue
-		}
-
-		// Split by pipe to handle pipelines on a single line
-		parts := strings.Split(trimmed, "|")
-		col := 0
-		for _, part := range parts {
-			exp := strings.TrimSpace(part)
-			if exp == "" {
-				col += len(part) + 1
-				continue
-			}
-
-			expType := ""
-			switch {
-			case strings.HasPrefix(exp, "subgraph"):
-				expType = "subgraph"
-			case strings.HasPrefix(exp, "from"):
-				expType = "from"
-			case strings.HasPrefix(exp, "make"):
-				expType = "make"
-			case strings.HasPrefix(exp, "remove"):
-				expType = "remove"
-			case strings.HasPrefix(exp, "collapse"):
-				expType = "collapse"
-			}
-
-			if expType != "" {
-				symbols = append(symbols, protocol.DocumentSymbol{
-					Name:   expType,
-					Detail: &exp,
-					Kind:   protocol.SymbolKindFunction,
-					Range: protocol.Range{
-						Start: protocol.Position{Line: uint32(i), Character: uint32(col)},
-						End:   protocol.Position{Line: uint32(i), Character: uint32(col + len(exp))},
-					},
-					SelectionRange: protocol.Range{
-						Start: protocol.Position{Line: uint32(i), Character: uint32(col)},
-						End:   protocol.Position{Line: uint32(i), Character: uint32(col + len(expType))},
-					},
-				})
-			}
-			col += len(part) + 1
-		}
-	}
-	return symbols
-}
-
-func gqlSemanticTokens(content string) []uint32 {
-	type gqlTok struct {
-		line    int
-		col     int
-		len     int
-		typeIdx uint32
-	}
-	var tokens []gqlTok
-
-	lines := strings.Split(content, "\n")
-	for lineIdx, line := range lines {
-		i := 0
-		for i < len(line) {
-			if line[i] == ' ' || line[i] == '\t' {
-				i++
-				continue
-			}
-
-			if line[i] == '#' {
-				break
-			}
-
-			if line[i] == '"' {
-				end := i + 1
-				for end < len(line) && line[end] != '"' {
-					end++
-				}
-				if end < len(line) {
-					end++
-				}
-				tokens = append(tokens, gqlTok{lineIdx, i, end - i, 1})
-				i = end
-				continue
-			}
-
-			if line[i] >= '0' && line[i] <= '9' || line[i] == '-' && i+1 < len(line) && line[i+1] >= '0' && line[i+1] <= '9' {
-				end := i
-				if line[end] == '-' {
-					end++
-				}
-				for end < len(line) && (line[end] >= '0' && line[end] <= '9' || line[end] == '.') {
-					end++
-				}
-				tokens = append(tokens, gqlTok{lineIdx, i, end - i, 2})
-				i = end
-				continue
-			}
-
-			if (line[i] >= 'a' && line[i] <= 'z') || (line[i] >= 'A' && line[i] <= 'Z') || line[i] == '_' {
-				end := i
-				for end < len(line) && ((line[end] >= 'a' && line[end] <= 'z') || (line[end] >= 'A' && line[end] <= 'Z') || (line[end] >= '0' && line[end] <= '9') || line[end] == '_') {
-					end++
-				}
-				word := line[i:end]
-				ti := uint32(0)
-				for _, kw := range gqlKeywords {
-					if kw.label == word {
-						ti = 0
-						break
-					}
-				}
-				if word == "true" || word == "false" {
-					ti = 3
-				}
-				tokens = append(tokens, gqlTok{lineIdx, i, end - i, ti})
-				i = end
-				continue
-			}
-
-			switch line[i] {
-			case '|', '@', '.', '(', ')', '+', '&', '-', '^':
-				tokens = append(tokens, gqlTok{lineIdx, i, 1, 6})
-			case '=', '!':
-				if i+1 < len(line) && line[i+1] == '=' {
-					tokens = append(tokens, gqlTok{lineIdx, i, 2, 6})
-					i += 2
-					continue
-				}
-				tokens = append(tokens, gqlTok{lineIdx, i, 1, 6})
-			}
-			i++
-		}
-	}
-
-	if len(tokens) == 0 {
-		return []uint32{}
-	}
-
-	var data []uint32
-	prevLine := uint32(0)
-	prevChar := uint32(0)
-	for _, tok := range tokens {
-		deltaLine := uint32(tok.line) - prevLine
-		deltaChar := uint32(tok.col)
-		if deltaLine == 0 {
-			deltaChar = uint32(tok.col) - prevChar
-		}
-		data = append(data, deltaLine, deltaChar, uint32(tok.len), tok.typeIdx, 0)
-		prevLine = uint32(tok.line)
-		prevChar = uint32(tok.col)
-	}
-	return data
-}
-
-func wordAt(content string, line, character int) *wordAtPos {
-	lines := strings.Split(content, "\n")
-	if line >= len(lines) {
-		return nil
-	}
-	lineStr := lines[line]
-	if character > len(lineStr) {
-		character = len(lineStr)
-	}
-
-	start := character
-	for start > 0 && isIdentRune(rune(lineStr[start-1])) {
-		start--
-	}
-	end := character
-	for end < len(lineStr) && isIdentRune(rune(lineStr[end])) {
-		end++
-	}
-
-	if start >= end {
-		if character < len(lineStr) && lineStr[character] == '@' {
-			return &wordAtPos{word: "@", line: line, colS: character, colE: character + 1}
-		}
-		return nil
-	}
-
-	return &wordAtPos{
-		word: lineStr[start:end],
-		line: line,
-		colS: start,
-		colE: end,
-	}
-}
-
-func isIdentRune(r rune) bool {
-	return (r >= 'A' && r <= 'Z') || (r >= 'a' && r <= 'z') || (r >= '0' && r <= '9') || r == '_'
-}
-
-func tokenize(content string) []gsl.Token {
-	l, err := gsl.NewLexer(strings.NewReader(content))
-	if err != nil {
-		return nil
-	}
-	return l.Tokenize()
-}
-
-func lineColToPos(tok gsl.Token) protocol.Position {
-	return protocol.Position{
-		Line:      uint32(max(0, tok.Line-1)),
-		Character: uint32(max(0, tok.Column-1)),
-	}
-}
-
-func (s *Server) complete(doc *document, line, character int) []protocol.CompletionItem {
-	if doc == nil {
-		return nil
-	}
-
-	lines := strings.Split(doc.content, "\n")
-	if line >= len(lines) {
-		return nil
-	}
-
-	lineStr := lines[line]
-	prefix := lineStr[:min(character, len(lineStr))]
-	trimmed := strings.TrimSpace(prefix)
-
-	tokens := tokenize(doc.content)
-	ctx := contextAt(tokens, line, character)
-
-	switch ctx {
-	case contextKeyword:
-		return keywordCompletions(trimmed)
-	case contextNodeName:
-		return nodeNameCompletions(doc)
-	case contextSetName:
-		return setNameCompletions(doc)
-	case contextAttribute:
-		return attributeCompletions(trimmed)
-	case contextEdgeTarget:
-		return nodeNameCompletions(doc)
-	case contextValue:
-		return valueCompletions(trimmed)
-	default:
-		return keywordCompletions(trimmed)
-	}
-}
-
-type completionContext int
-
-const (
-	contextUnknown completionContext = iota
-	contextKeyword
-	contextNodeName
-	contextSetName
-	contextAttribute
-	contextEdgeTarget
-	contextValue
-)
-
-func contextAt(tokens []gsl.Token, line, character int) completionContext {
-	var bestTok *gsl.Token
-	bestIdx := -1
-
-	for i, tok := range tokens {
-		if tok.Type == gsl.TOKEN_EOF {
-			continue
-		}
-		if int(tok.Line-1) != line {
-			continue
-		}
-		if int(tok.Column-1) > character {
-			continue
-		}
-		if bestTok == nil || int(tok.Column-1) >= int(bestTok.Column-1) {
-			t := tok
-			bestTok = &t
-			bestIdx = i
-		}
-	}
-
-	if bestTok == nil {
-		return contextKeyword
-	}
-
-	switch bestTok.Type {
-	case gsl.TOKEN_AT:
-		return contextSetName
-	case gsl.TOKEN_LBRACKET:
-		return contextAttribute
-	case gsl.TOKEN_COMMA, gsl.TOKEN_ARROW:
-		return contextEdgeTarget
-	case gsl.TOKEN_EQUALS:
-		return contextValue
-	case gsl.TOKEN_NODE:
-		return contextNodeName
-	case gsl.TOKEN_SET:
-		return contextNodeName
-	case gsl.TOKEN_COLON:
-		if bestIdx > 0 && tokens[bestIdx-1].Type == gsl.TOKEN_IDENT {
-			return contextEdgeTarget
-		}
-	case gsl.TOKEN_LBRACE:
-		if bestIdx > 0 {
-			prev := tokens[bestIdx-1]
-			if prev.Type == gsl.TOKEN_RBRACKET || prev.Type == gsl.TOKEN_IDENT || prev.Type == gsl.TOKEN_STRING {
-				return contextKeyword
-			}
-		}
-	}
-
-	return contextKeyword
-}
-
-func keywordCompletions(prefix string) []protocol.CompletionItem {
-	keywords := []struct {
-		label  string
-		detail string
-		kind   protocol.CompletionItemKind
-	}{
-		{"node", "Declare a node", protocol.CompletionItemKindKeyword},
-		{"set", "Declare a set", protocol.CompletionItemKindKeyword},
-		{"true", "Boolean true", protocol.CompletionItemKindKeyword},
-		{"false", "Boolean false", protocol.CompletionItemKindKeyword},
-	}
-
-	var items []protocol.CompletionItem
-	for _, kw := range keywords {
-		if prefix == "" || strings.HasPrefix(kw.label, prefix) {
-			items = append(items, protocol.CompletionItem{
-				Label:      kw.label,
-				Detail:     protocol.NewOptional(kw.detail),
-				Kind:       kw.kind,
-				InsertText: protocol.NewOptional(kw.label),
-			})
-		}
-	}
-	return items
-}
-
-func nodeNameCompletions(doc *document) []protocol.CompletionItem {
-	names := make(map[string]struct{})
-
-	if doc.graph != nil {
-		for name := range doc.graph.GetNodes() {
-			names[name] = struct{}{}
-		}
-	}
-
-	extractNodeNames(doc.content, names)
-
-	var items []protocol.CompletionItem
-	for name := range names {
-		items = append(items, protocol.CompletionItem{
-			Label:  name,
-			Kind:   protocol.CompletionItemKindVariable,
-			Detail: protocol.NewOptional("node"),
-		})
-	}
-	sort.Slice(items, func(i, j int) bool { return items[i].Label < items[j].Label })
-	return items
-}
-
-func extractNodeNames(content string, names map[string]struct{}) {
-	lines := strings.Split(content, "\n")
-	for _, line := range lines {
-		trimmed := strings.TrimSpace(line)
-		if strings.HasPrefix(trimmed, "node ") {
-			parts := strings.Fields(trimmed)
-			if len(parts) >= 2 {
-				name := parts[1]
-				if strings.HasPrefix(name, "[") || strings.HasPrefix(name, ":") {
-					continue
-				}
-				names[name] = struct{}{}
-			}
-		}
-		// Extract node names from edge declarations: "A -> B" or "A, B -> C"
-		if strings.Contains(trimmed, "->") {
-			fields := strings.Fields(trimmed)
-			for _, f := range fields {
-				if f == "->" {
-					continue
-				}
-				if strings.HasPrefix(f, "@") || strings.HasPrefix(f, "[") || f == ":" {
-					continue
-				}
-				// Clean trailing punctuation
-				f = strings.TrimRight(f, ",;:")
-				if isIdent(f) && !isKeyword(f) {
-					names[f] = struct{}{}
-				}
-			}
-		}
-	}
-}
-
-func isIdent(s string) bool {
-	if s == "" {
-		return false
-	}
-	for _, r := range s {
-		if !((r >= 'A' && r <= 'Z') || (r >= 'a' && r <= 'z') || (r >= '0' && r <= '9') || r == '_') {
-			return false
-		}
-	}
-	return true
-}
-
-func isKeyword(s string) bool {
-	switch s {
-	case "node", "set", "true", "false":
-		return true
-	}
-	return false
-}
-
-func setNameCompletions(doc *document) []protocol.CompletionItem {
-	names := make(map[string]struct{})
-
-	if doc.graph != nil {
-		for name := range doc.graph.GetSets() {
-			names[name] = struct{}{}
-		}
-	}
-
-	var items []protocol.CompletionItem
-	for name := range names {
-		items = append(items, protocol.CompletionItem{
-			Label:  name,
-			Kind:   protocol.CompletionItemKindVariable,
-			Detail: protocol.NewOptional("set"),
-		})
-	}
-	sort.Slice(items, func(i, j int) bool { return items[i].Label < items[j].Label })
-	return items
-}
-
-func attributeCompletions(prefix string) []protocol.CompletionItem {
-	known := []struct {
-		label  string
-		detail string
-		kind   protocol.CompletionItemKind
-	}{
-		{"text", "Display text", protocol.CompletionItemKindProperty},
-		{"parent", "Parent/reference", protocol.CompletionItemKindProperty},
-		{"color", "Color attribute", protocol.CompletionItemKindProperty},
-		{"weight", "Weight attribute", protocol.CompletionItemKindProperty},
-		{"style", "Style attribute", protocol.CompletionItemKindProperty},
-		{"shape", "Shape attribute", protocol.CompletionItemKindProperty},
-	}
-
-	var items []protocol.CompletionItem
-	for _, a := range known {
-		if prefix == "" || strings.HasPrefix(a.label, prefix) {
-			items = append(items, protocol.CompletionItem{
-				Label:  a.label,
-				Kind:   a.kind,
-				Detail: protocol.NewOptional(a.detail),
-			})
-		}
-	}
-	return items
-}
-
-func valueCompletions(prefix string) []protocol.CompletionItem {
-	var items []protocol.CompletionItem
-	if prefix == "" || strings.HasPrefix("true", prefix) {
-		items = append(items, protocol.CompletionItem{
-			Label:  "true",
-			Kind:   protocol.CompletionItemKindKeyword,
-			Detail: protocol.NewOptional("boolean"),
-		})
-	}
-	if prefix == "" || strings.HasPrefix("false", prefix) {
-		items = append(items, protocol.CompletionItem{
-			Label:  "false",
-			Kind:   protocol.CompletionItemKindKeyword,
-			Detail: protocol.NewOptional("boolean"),
-		})
-	}
-	return items
-}
-
-func (s *Server) hoverFor(graph *gsl.Graph, word string) string {
-	if graph == nil {
-		return ""
-	}
-
-	var parts []string
-
-	if node := graph.GetNode(word); node != nil {
-		parts = append(parts, fmt.Sprintf("**Node** `%s`", node.ID))
-		if len(node.Attributes) > 0 {
-			parts = append(parts, "\n\n**Attributes:**")
-			keys := make([]string, 0, len(node.Attributes))
-			for k := range node.Attributes {
-				keys = append(keys, k)
-			}
-			sort.Strings(keys)
-			for _, k := range keys {
-				parts = append(parts, fmt.Sprintf("\n- `%s` = `%v`", k, node.Attributes[k]))
-			}
-		}
-		if len(node.Sets) > 0 {
-			setNames := make([]string, 0, len(node.Sets))
-			for s := range node.Sets {
-				setNames = append(setNames, s)
-			}
-			sort.Strings(setNames)
-			parts = append(parts, "\n\n**Sets:**"+fmt.Sprintf(" @%s", strings.Join(setNames, ", @")))
-		}
-	}
-
-	if sets := graph.GetSets(); sets != nil {
-		if s, ok := sets[word]; ok {
-			if len(parts) > 0 {
-				parts = append(parts, "\n---\n")
-			}
-			parts = append(parts, fmt.Sprintf("**Set** `%s`", s.ID))
-			if len(s.Attributes) > 0 {
-				parts = append(parts, "\n\n**Attributes:**")
-				keys := make([]string, 0, len(s.Attributes))
-				for k := range s.Attributes {
-					keys = append(keys, k)
-				}
-				sort.Strings(keys)
-				for _, k := range keys {
-					parts = append(parts, fmt.Sprintf("\n- `%s` = `%v`", k, s.Attributes[k]))
-				}
-			}
-		}
-	}
-
-	for _, e := range graph.GetEdges() {
-		if word == e.From || word == e.To || (e.Label != "" && word == e.Label) {
-			if len(parts) > 0 {
-				parts = append(parts, "\n---\n")
-			}
-			parts = append(parts, fmt.Sprintf("**Edge** `%s` → `%s`", e.From, e.To))
-			if e.Label != "" {
-				parts = append(parts, fmt.Sprintf("\n\nLabel: `%s`", e.Label))
-			}
-			if len(e.Attributes) > 0 {
-				parts = append(parts, "\n\n**Attributes:**")
-				keys := make([]string, 0, len(e.Attributes))
-				for k := range e.Attributes {
-					keys = append(keys, k)
-				}
-				sort.Strings(keys)
-				for _, k := range keys {
-					parts = append(parts, fmt.Sprintf("\n- `%s` = `%v`", k, e.Attributes[k]))
-				}
-			}
-			break
-		}
-	}
-
-	if len(parts) == 0 {
-		return ""
-	}
-	return strings.Join(parts, "")
-}
-
-func findDeclPosition(tokens []gsl.Token, name string, docURI uri.URI) *protocol.Location {
-	for i, tok := range tokens {
-		if tok.Literal != name || tok.Type != gsl.TOKEN_IDENT {
-			continue
-		}
-		if i >= 1 {
-			prev := tokens[i-1]
-			if prev.Type == gsl.TOKEN_NODE || prev.Type == gsl.TOKEN_SET {
-				pos := lineColToPos(tok)
-				return &protocol.Location{
-					URI: docURI,
-					Range: protocol.Range{
-						Start: pos,
-						End:   pos,
-					},
-				}
-			}
-		}
-	}
+func (s *Server) DidChangeWatchedFiles(ctx context.Context, params *protocol.DidChangeWatchedFilesParams) error {
 	return nil
 }
-
-func findReferences(tokens []gsl.Token, name string, docURI uri.URI) []protocol.Location {
-	var locs []protocol.Location
-	for _, tok := range tokens {
-		if tok.Literal == name && tok.Type == gsl.TOKEN_IDENT {
-			pos := lineColToPos(tok)
-			locs = append(locs, protocol.Location{
-				URI: docURI,
-				Range: protocol.Range{
-					Start: pos,
-					End:   pos,
-				},
-			})
-		}
-	}
-	return locs
-}
-
-func buildSymbols(tokens []gsl.Token) []protocol.DocumentSymbol {
-	var symbols []protocol.DocumentSymbol
-
-	for i, tok := range tokens {
-		if tok.Type == gsl.TOKEN_EOF {
-			break
-		}
-
-		if tok.Type == gsl.TOKEN_NODE && i+1 < len(tokens) && tokens[i+1].Type == gsl.TOKEN_IDENT {
-			nameTok := tokens[i+1]
-			endPos := lineColToPos(nameTok)
-			endPos.Character += uint32(len(nameTok.Literal))
-
-			symbols = append(symbols, protocol.DocumentSymbol{
-				Name:           nameTok.Literal,
-				Kind:           protocol.SymbolKindVariable,
-				Range:          protocol.Range{Start: lineColToPos(tok), End: endPos},
-				SelectionRange: protocol.Range{Start: lineColToPos(nameTok), End: endPos},
-			})
-		}
-
-		if tok.Type == gsl.TOKEN_SET && i+1 < len(tokens) && tokens[i+1].Type == gsl.TOKEN_IDENT {
-			nameTok := tokens[i+1]
-			endPos := lineColToPos(nameTok)
-			endPos.Character += uint32(len(nameTok.Literal))
-
-			symbols = append(symbols, protocol.DocumentSymbol{
-				Name:           nameTok.Literal,
-				Kind:           protocol.SymbolKindNamespace,
-				Range:          protocol.Range{Start: lineColToPos(tok), End: endPos},
-				SelectionRange: protocol.Range{Start: lineColToPos(nameTok), End: endPos},
-			})
-		}
-	}
-
-	return symbols
-}
-
-func semanticTokenTypes() []string {
-	return []string{
-		"keyword",
-		"string",
-		"number",
-		"boolean",
-		"type",
-		"property",
-		"operator",
-		"comment",
-	}
-}
-
-func tokenTypeToSemanticIndex(tt gsl.TokenType) uint32 {
-	switch tt {
-	case gsl.TOKEN_NODE, gsl.TOKEN_SET:
-		return 0
-	case gsl.TOKEN_TRUE, gsl.TOKEN_FALSE:
-		return 3
-	case gsl.TOKEN_STRING:
-		return 1
-	case gsl.TOKEN_NUMBER:
-		return 2
-	case gsl.TOKEN_IDENT:
-		return 4
-	case gsl.TOKEN_EQUALS, gsl.TOKEN_ARROW:
-		return 6
-	case gsl.TOKEN_AT, gsl.TOKEN_COLON, gsl.TOKEN_COMMA:
-		return 6
-	case gsl.TOKEN_LBRACKET, gsl.TOKEN_RBRACKET, gsl.TOKEN_LBRACE, gsl.TOKEN_RBRACE:
-		return 6
-	default:
-		return 0
-	}
-}
-
-func encodeSemanticTokens(tokens []gsl.Token) []uint32 {
-	if len(tokens) == 0 {
-		return []uint32{}
-	}
-
-	var data []uint32
-	prevLine := uint32(0)
-	prevChar := uint32(0)
-
-	hasCommentPrefix := false
-	commentLine := uint32(0)
-	commentChar := uint32(0)
-
-	for _, tok := range tokens {
-		if tok.Type == gsl.TOKEN_EOF {
-			break
-		}
-
-		pos := lineColToPos(tok)
-
-		if tok.Type == gsl.TOKEN_ILLEGAL {
-			if tok.Literal == "#" {
-				hasCommentPrefix = true
-				commentLine = pos.Line
-				commentChar = pos.Character
-				continue
-			}
-			if hasCommentPrefix && pos.Line == commentLine && pos.Character >= commentChar {
-				continue
-			}
-			hasCommentPrefix = false
-		} else {
-			if hasCommentPrefix {
-				if pos.Line == commentLine {
-					continue
-				}
-				hasCommentPrefix = false
-			}
-		}
-
-		tokLen := uint32(len(tok.Literal))
-
-		typeIdx := tokenTypeToSemanticIndex(tok.Type)
-
-		deltaLine := pos.Line - prevLine
-		deltaChar := pos.Character
-		if deltaLine == 0 {
-			deltaChar = pos.Character - prevChar
-		}
-
-		data = append(data, deltaLine, deltaChar, tokLen, typeIdx, 0)
-		prevLine = pos.Line
-		prevChar = pos.Character
-	}
-
-	return data
-}
-
-var (
-	_ = regexp.Compile
-)
